pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-agent"
version = "0.1.0"
description = "GenAI Image Generation Agent powered by LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "tavily-python>=0.5.0",
    "openai>=1.60.0",
    "huggingface-hub>=0.25.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
]

[project.scripts]
image-agent = "image_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/image_agent"]
